TST: split series/test_indexing.py (#18614) (#20006)
